feat : team-pig #57 - 삭제 confirm 적용

+++ src/feature/room/RoomCard.jsx
@@ -76,9 +76,9 @@ const RoomCard = ({
   const deleteRoom = (e) => {
     e.stopPropagation();
     setIsDisplayDrop(false);
-    if (userId === master) {
+    const deleteCheck = window.confirm("🗑정말 이 방을 삭제할까요?");
+    if (userId === master && deleteCheck) {
       dispatch(__deleteRoom(roomId));
-      window.alert("🗑정말 이 방을 삭제할까요?");
     }
   };
